refact: 랭킹 함수 구현

+++ src/main/java/lotto/Lotto.java
@@ -1,10 +1,7 @@
 package lotto;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collector;
 import java.util.stream.Collectors;
 
 public class Lotto {
@@ -33,7 +30,7 @@ public class Lotto {
     }
 
     public int countMatchingWinningLotto(Lotto lotto) {
-        return lotto.numbers.stream()  
+        return (int) lotto.numbers.stream()  
             .filter(numbers::contains)  
             .count();  
     }
